Beams

ABCDMatrixOptics.AbstractBeam — Type

abstract type AbstractBeam{T}

Serves as abstract type for GeometricBeam and GaussianBeam.

Simple Geometric Beam

ABCDMatrixOptics.GeometricBeam — Type

GeometricBeam(w=0.0, k=0.0, z=0.0)

Defines a geometrical ray beam with keywords. Distance to the axis is w, the angle with respect to the axis is k. zpos is the initial position along the optical axis.

See also GaussianBeam.

Example

julia> GeometricBeam(w=1.0)
GeometricBeam{Float64}(1.0, 0.0, 0.0)

Simple Gaussian Beam

ABCDMatrixOptics.GaussianBeam — Type

struct GaussianBeam{T} <: AbstractBeam{T}
    q
    zpos
    n
    λ
end

ABCDMatrixOptics.z — Function

z(beam::GaussianBeam)

Return the distance to the beam waist z of the Gaussian beam.

See also z,zR,w,R,w0.

ABCDMatrixOptics.zR — Function

z(beam::GaussianBeam)

Return Rayleigh length zR of the Gaussian beam.

See also z, w, R, w0.

ABCDMatrixOptics.R — Function

R(beam::GaussianBeam)

Return curvature R of the Gaussian beam.

See also z, zR, w, w0.

ABCDMatrixOptics.w0 — Function

w0(beam::GaussianBeam)

Return w0 of the Gaussian beam.

See also z, zR, w, R.

ABCDMatrixOptics.w — Function

w(beam::GaussianBeam)

Return current width w of the Gaussian beam.

See also z, zR, R, w0.