Elements

Elements or Vector{<:Elements} can be also compare with ≈.

ABCDMatrixOptics.FreeSpace — Type

FreeSpace(dz)

Construct a free space propagation over distance dz.

ABCDMatrixOptics.ThinLens — Type

ThinLens(f)

Creates a thin lens with focal length f.

ABCDMatrixOptics.ThickLens — Type

ThickLens(; n_lens, R1, R2, t)

Construct a thick lens with the keywords:

R1 radius of curvature of first surface

R2 radius of curvature of second surface

t: thickness of lens

n_lens=1.5 refractive index of lens

n1=1: refractive index of the medium of the incoming side

n2=1: refractive index of the medium of the exiting side

ABCDMatrixOptics.Interface — Type

Interface(n1, n2)

Creates a flat interface with refractive index n1 on the entering side and n2 on the new medium.

Interface(n1, n2, R)

Creates a curved interface with radius R and with refractive index n1 on the entering side and n2 on the new medium.

ABCDMatrixOptics.dz — Function

dz(element::Element)

Returns how much an element changes the optical distance z.